# api-spec.yaml
openapi: 3.1.0
info:
  title: flecs-core
  description: ''
  contact:
    name: FLECS Technologies GmbH
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
servers:
- url: http://{address}/{version}
  variables:
    address:
      default: localhost
      description: IP Address of FLECS Daemon
    version:
      default: v2
      description: API Version
paths:
...
  /instances/{instance_id}/provides:
    get:
      tags:
      - Experimental
      description: Get all provided features of the specified instance
      operationId: get_instances_{instance_id}_provides
      parameters:
      - name: instance_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/HexString8'
      responses:
        '200':
          description: Information for all features and their config provided by this instance
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/FeatureInfo'
                propertyNames:
                  type: string
                  minLength: 1
                  pattern: ^([\w.-]+)$
        '404':
          description: Instance id not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
  /instances/{instance_id}/provides/{feature}:
    get:
      tags:
      - Experimental
      description: Check if the specified instance provides the specified feature
      operationId: get_instances_{instance_id}_provides_{feature}
      parameters:
      - name: instance_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/HexString8'
      - name: feature
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/FeatureKey'
      responses:
        '200':
          description: Information of the specified feature provided by the specified instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureInfo'
        '404':
          description: Instance not found or feature not provided by instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceNotFoundOrFeatureNotProvided'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
...
  /providers/auth/core:
    get:
      tags:
      - Experimental
      description: Get information on the core auth provider
      operationId: get_providers_auth_core
      responses:
        '204':
          description: How the core auth provider is currently set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderReference'
        '404':
          description: No core auth provider set
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
...
  /providers/auth/default:
    get:
      tags:
      - Experimental
      description: Get the default auth provider
      operationId: get_providers_auth_default
      responses:
        '200':
          description: Default auth provider was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProvider'
        '404':
          description: No default auth provider was found
    put:
      tags:
      - Experimental
      description: Set the default auth provider
      operationId: put_providers_auth_default
      requestBody:
        description: Id of the provider that should be used as the default auth provider
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutDefaultProviderRequest'
        required: true
      responses:
        '200':
          description: Default auth provider was replaced
        '201':
          description: Default auth provider was set
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
    delete:
      tags:
      - Experimental
      description: Remove the default auth provider
      operationId: delete_providers_auth_default
      responses:
        '200':
          description: Default auth provider was removed
        '404':
          description: No default auth provider was found
        '409':
          description: The current state does not allow the removal of the default auth provider, e.g. a running instance is using it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalInfo'
...
  /providers/auth/{id}:
    get:
      tags:
      - Experimental
      description: Get the auth provider with the specified id
      operationId: get_providers_auth_{id}
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/HexString8'
      responses:
        '200':
          description: Auth provider was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProvider'
        '404':
          description: Auth provider was not found
...
components:
  schemas:
...
    AdditionalInfo:
      type: object
      required:
      - additional_info
      properties:
        additional_info:
          type: string
...
    AuthProvider:
      type: object
      required:
      - issuer_url
      - name
      - kind
      - config
      properties:
        config: {}
        issuer_url:
          type: string
          format: uri
        kind:
          type: string
        name:
          type: string
...
    FeatureInfo:
      type: object
      required:
      - config
      properties:
        config: {}
    FeatureKey:
      type: string
      minLength: 1
      pattern: ^([\w.-]+)$
...
    HexString8:
      type: string
      maxLength: 8
      minLength: 8
      pattern: ^[0-9a-fA-F]{8}$
    InstanceNotFoundOrFeatureNotProvided:
      type: string
      enum:
      - FeatureNotProvided
      - InstanceNotFound
...
    ProviderReference:
      oneOf:
      - type: string
        enum:
        - Default
      - type: object
        required:
        - Provider
        properties:
          Provider:
            $ref: '#/components/schemas/HexString8'
    PutDefaultProviderRequest:
      type: object
      required:
      - provider_id
      properties:
        provider_id:
          $ref: '#/components/schemas/HexString8'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []